+++ test/blackbox-tests/test-cases/misc/run.t
@@ -1,4 +1,6 @@
   $ $JBUILDER runtest -j1 --root .
+  File "jbuild", line 54, characters 10-209:
+  Warning: Directory _build/default/dir-that-doesnt-exist doesn't exist.
           diff alias runtest
           diff alias runtest
           diff alias runtest